// js/src/common/types.ts
export interface RequestRoute {
  name: string | null;
  handler: string;
  parameters: Record<string, string>;
  internal: boolean;
}

export interface RequestActor {
  id?: number | null;
  username?: string | null;
  isGuest?: boolean;
  isAdmin?: boolean;
  groups?: string[];
  authentication: string;
}

export interface RequestData {
  method: string;
  uri: string;
  status: number;
  route: RequestRoute;
  actor: RequestActor;
  query: Record<string, string>;
  jsonApi: Record<string, string>;
  requestHeaders: Record<string, string>;
  responseHeaders: Record<string, string>;
}